fix breaking changes after rebase with main
